add id/created fields to the report JSON as well
This is important data to send back when people are getting the reports,
and it's easier than adding it everywhere it needs to be. We may, in the
future, change the `created` and `id` columns to be generated from the
JSON report, but SQLite doesn't do this which makes it more difficult.