allow test reports to be reprocessed
If a test report has already been processed based on looking up its guid
in the stats table, the Stats entry will be updated instead of a new one
created.

To do this well, I needed to change the name of the "guid" unique
constraint on the stats table, so that perhaps later someday we can
change the name of the table itself without breaking the API for
everyone.