fix issue deserializing json with control chars
MySQL seems to be giving us back the JSON as a binary blob. This is
causing a problem when the JSON contains ASCII control characters
(U+0000-U+001F). Since these must be escaped in JSON (as `\u####`), no
standards-compliant JSON parser will decode them.

Now we pre-escape them to allow the JSON parser to parse it correctly.