Fix Perl version testing config in Travis
In order to ensure that the environment variables are used in the builds
for Perl 5.24 and blead, one must define the Perl versions within the
`include` section of the `matrix` directive.  This ensures that instead
of running two builds using Perl 5.24 (one without environment
variables, and one with) that only one build with the required
environment variables is run.