make `created` be only set from us
Metabase reports come with a "creation" time. Some of the test reporters
are sending us reports from hours or days in the past.

Since we use this created time to create feeds, we need to know when we
received the report, not when it was created on the user's machine. That
probably means we should rename this column to "reported" or "submitted"
or "received", but for now this will be fine

Fixes #19