add CPAN::Testers::ReportsDir module
This is a new, plain-text storage for CPAN Testers reports. The plan is
to convert all the existing Metabase-style reports in the database and
instead keep them as plain YAML files using this module instead. Then,
CPAN::Testers::Schema can look in this for reports before going to
a database to do so.